refactor: modify the 9/6 codereview companies section
