+++ apps/companies/forms/company_form.py
@@ -11,20 +11,18 @@ class CompanyForm(ModelForm):
             "tel",
             "url",
             "address",
-            "describe",
-            "total_headcount",
+            "description",
+            "employees",
             "name",
             "email",
-            "owner_tel",
         ]
         labels = {
             "title": "公司名稱",
             "tel": "公司電話",
             "url": "公司網址",
             "address": "公司地址",
-            "describe": "公司描述",
-            "total_headcount": "員工人數",
+            "description": "公司描述",
+            "employees": "員工人數",
             "name": "負責人姓名",
             "email": "負責人Email",
-            "owner_tel": "負責人電話",
         }
