+++ apps/companies/views.py
@@ -1,32 +1,27 @@
 from django.shortcuts import get_object_or_404, redirect, render
+from django.views.decorators.http import require_POST
 
-from .forms.form import CompanyForm
+from .forms.company_form import CompanyForm
 from .models import Company
 
 
 def index(request):
-
     if request.method == "POST":
         form = CompanyForm(request.POST)
         if form.is_valid():
             form.save()
             return redirect("companies:index")
-
     companies = Company.objects.order_by("-id")
-
     return render(request, "companies/index.html", {"companies": companies})
-
 
 def new(request):
     form = CompanyForm()
     return render(request, "companies/new.html", {"form": form})
 
-
 def edit(request, id):
     company = get_object_or_404(Company, id=id)
     form = CompanyForm(instance=company)
     return render(request, "companies/edit.html", {"form": form, "company": company})
-
 
 def show(request, id):
     if request.method == "POST":
@@ -43,9 +38,8 @@ def show(request, id):
     company = get_object_or_404(Company, id=id)
     return render(request, "companies/show.html", {"company": company})
 
-
+@require_POST
 def delete(request, id):
     company = get_object_or_404(Company, id=id)
-    if request.method == "POST":
-        company.mark_delete()
-        return redirect("companies:index")
+    company.mark_delete()
+    return redirect("companies:index")
